Move deprecated methods to the deprecated module
