+++ src/lib.rs
@@ -92,11 +92,15 @@ pub use timeunits::*;
 mod timeseries;
 pub use timeseries::*;
 
+/// This module defines all of the deprecated methods. To see if you're using any deprecated features, build hifitime without the `deprecated` feature flag.
+#[cfg(feature = "deprecated")]
+mod deprecated;
+
 #[allow(deprecated)]
 pub mod prelude {
     pub use crate::{
-        Duration, Epoch, Errors, Freq, Frequencies, TimeScale, TimeSeries, TimeSystem, TimeUnits,
-        Unit,
+        deprecated::TimeSystem, Duration, Epoch, Errors, Freq, Frequencies, TimeScale, TimeSeries,
+        TimeUnits, Unit,
     };
 }
